requirement.txt
python==3.8
streamlit
openai
pinecone-client